RF: moving qbx_merge to segment.clustering module
