+++ dipy/tracking/streamline.py
@@ -653,3 +653,15 @@ def values_from_volume(data, streamlines, affine=None):
         return _extract_vals(data, streamlines, affine=affine)
     else:
         raise ValueError("Data needs to have 3 or 4 dimensions")
+
+
+def check_range(streamline, gt, lt):
+    length_s = length(streamline)
+    if (length_s > gt) & (length_s < lt):
+        return True
+    else:
+        return False
+
+
+def nbytes(streamlines):
+    return streamlines._data.nbytes / 1024. ** 2
